Use `Fields` wrapper struct directly, and move to own module
